Update quiz.py
Fixed the functions so that the script doesn't rely on global variables. Rather, I store quizlet's data in a single dictionary, and then use an xrange() function to loop through the terms, printing out their mneumonics with printAll()... I will admit, it was a bit bold using printAll() without even storing the information in a cache-esque list first, but you have to be bold sometimes. It's just the way of the samurai, if you will.